Revert subpath rename and tighten the deprecation surface
Addresses the second round of review feedback.

- Keep `experimental/markdown-renderer` and `experimental/tiff-renderer` as
  the only import paths, now carrying the new component names. They get
  renamed when the `experimental/` prefix is dropped, so consumers change
  import paths once rather than twice. The shims added in the previous commit
  are merged into these barrels, so each public subpath maps to one source
  file named after it, and the aggregate `experimental` barrel exposes both
  naming generations without ambiguous star exports.
- Derive `PdfViewerInstanceOptions` from `BasePdfViewerProps` instead of
  restating it, so the two cannot drift again. They already had: the options
  type spelled the highlight flag `highlightEnabled` and omitted
  `downloadFileName` entirely. `enableHighlight` is now the spelling, with
  `highlightEnabled` kept as a deprecated alias, and `downloadFileName` stays
  excluded because the hooks tier never implemented it.
- Rewrite the changeset as an old to new migration list.
- Drop the `MarkdownViewerMediaProps` and `TiffViewerMediaProps` aliases so
  removing the `Media` suffix is a uniform break across all eight viewers. It
  cannot be aliased for pdf, whose old `PdfViewerProps` name was reused for a
  different type, and aliasing seven of eight would be more confusing than
  none.
