+++ .monorepolint.config.mjs
@@ -583,13 +583,11 @@ const archetypeRules = archetypes(standardPackageRules, {
         "./experimental/filter-list",
         "./experimental/image-viewer",
         "./experimental/markdown-renderer",
-        "./experimental/markdown-viewer",
         "./experimental/object-table",
         "./experimental/pdf-viewer",
         "./experimental/spreadsheet-viewer",
         "./experimental/theme",
         "./experimental/tiff-renderer",
-        "./experimental/tiff-viewer",
         "./experimental/video-viewer",
         "./experimental/xml-viewer",
       ],
