+++ packages/react-components-storybook/src/stories/MarkdownViewer/MarkdownViewer.stories.tsx
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import type { BaseMarkdownViewerProps } from "@osdk/react-components/experimental/markdown-viewer";
-import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-viewer";
+import type { BaseMarkdownViewerProps } from "@osdk/react-components/experimental/markdown-renderer";
+import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-renderer";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
 const SAMPLE_MARKDOWN = `# Sample Document
@@ -39,7 +39,7 @@ This is a **sample markdown** document that demonstrates the *BaseMarkdownViewer
 ## Code Example
 
 \`\`\`typescript
-import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-viewer";
+import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-renderer";
 
 function App() {
   return <BaseMarkdownViewer content="# Hello World" />;
@@ -143,7 +143,7 @@ export const MinimalContent: Story = {
   parameters: {
     docs: {
       source: {
-        code: `import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-viewer";
+        code: `import { BaseMarkdownViewer } from "@osdk/react-components/experimental/markdown-renderer";
 
 <BaseMarkdownViewer content="# Hello World\\n\\nA simple paragraph with **bold** and *italic*." />`,
       },
