+++ packages/react-components/package.json
@@ -106,15 +106,6 @@
       "require": "./build/cjs/public/experimental/markdown-renderer.cjs",
       "default": "./build/browser/public/experimental/markdown-renderer.js"
     },
-    "./experimental/markdown-viewer": {
-      "browser": "./build/browser/public/experimental/markdown-viewer.js",
-      "import": {
-        "types": "./build/types/public/experimental/markdown-viewer.d.ts",
-        "default": "./build/esm/public/experimental/markdown-viewer.js"
-      },
-      "require": "./build/cjs/public/experimental/markdown-viewer.cjs",
-      "default": "./build/browser/public/experimental/markdown-viewer.js"
-    },
     "./experimental/object-table": {
       "browser": "./build/browser/public/experimental/object-table.js",
       "import": {
@@ -160,15 +151,6 @@
       "require": "./build/cjs/public/experimental/tiff-renderer.cjs",
       "default": "./build/browser/public/experimental/tiff-renderer.js"
     },
-    "./experimental/tiff-viewer": {
-      "browser": "./build/browser/public/experimental/tiff-viewer.js",
-      "import": {
-        "types": "./build/types/public/experimental/tiff-viewer.d.ts",
-        "default": "./build/esm/public/experimental/tiff-viewer.js"
-      },
-      "require": "./build/cjs/public/experimental/tiff-viewer.cjs",
-      "default": "./build/browser/public/experimental/tiff-viewer.js"
-    },
     "./experimental/video-viewer": {
       "browser": "./build/browser/public/experimental/video-viewer.js",
       "import": {
@@ -199,7 +181,7 @@
     }
   },
   "scripts": {
-    "check-attw": "attw --pack . --exclude-entrypoints ./styles.css ./experimental/action-form ./experimental/aip-agent-chat ./experimental/cbac-picker ./experimental/document-viewer ./experimental/email-viewer ./experimental/filter-list ./experimental/image-viewer ./experimental/markdown-renderer ./experimental/markdown-viewer ./experimental/object-table ./experimental/pdf-viewer ./experimental/spreadsheet-viewer ./experimental/theme ./experimental/tiff-renderer ./experimental/tiff-viewer ./experimental/video-viewer ./experimental/xml-viewer",
+    "check-attw": "attw --pack . --exclude-entrypoints ./styles.css ./experimental/action-form ./experimental/aip-agent-chat ./experimental/cbac-picker ./experimental/document-viewer ./experimental/email-viewer ./experimental/filter-list ./experimental/image-viewer ./experimental/markdown-renderer ./experimental/object-table ./experimental/pdf-viewer ./experimental/spreadsheet-viewer ./experimental/theme ./experimental/tiff-renderer ./experimental/video-viewer ./experimental/xml-viewer",
     "check-gen-props": "node scripts/gen-props.mjs --check",
     "check-spelling": "cspell --quiet .",
     "clean": "rm -rf lib dist types build tsconfig.tsbuildinfo",
