+++ packages/react-components/src/pdf-viewer/PdfViewerContext.tsx
@@ -146,7 +146,7 @@ export function usePdfViewerInstance(
   const { highlightModeActive, toggleHighlightMode } = usePdfHighlightMode({
     pdfViewerRef: viewer.pdfViewerRef,
     document: viewer.document,
-    enabled: options.highlightEnabled ?? false,
+    enabled: options.enableHighlight ?? options.highlightEnabled ?? false,
     onTextHighlight: options.onTextHighlight,
     onHighlightDelete: options.onHighlightDelete,
   });
@@ -163,7 +163,9 @@ export function usePdfViewerInstance(
   const annotations = options.annotations ?? EMPTY_ANNOTATION_ARRAY;
   const annotationsByPage = usePdfAnnotationsByPage(annotations);
 
-  const highlightEnabled = options.highlightEnabled ?? false;
+  // `highlightEnabled` is the deprecated spelling of `enableHighlight`.
+  const highlightEnabled =
+    options.enableHighlight ?? options.highlightEnabled ?? false;
   const enableDownload = options.enableDownload ?? false;
   const enableFormSave = options.onFormSubmit != null && hasFormFields;
   const { onAnnotationClick, outlineIcons } = options;
