+++ packages/react-components/src/public/experimental.ts
@@ -20,11 +20,11 @@ export * from "./experimental/document-viewer.js";
 export * from "./experimental/email-viewer.js";
 export * from "./experimental/filter-list.js";
 export * from "./experimental/image-viewer.js";
-export * from "./experimental/markdown-viewer.js";
+export * from "./experimental/markdown-renderer.js";
 export * from "./experimental/object-table.js";
 export * from "./experimental/pdf-viewer.js";
 export * from "./experimental/spreadsheet-viewer.js";
 export * from "./experimental/theme.js";
-export * from "./experimental/tiff-viewer.js";
+export * from "./experimental/tiff-renderer.js";
 export * from "./experimental/video-viewer.js";
 export * from "./experimental/xml-viewer.js";
