--- packages/react-components/src/public/experimental/markdown-viewer.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// BaseMarkdownViewer
-export { BaseMarkdownViewer } from "../../markdown-viewer/BaseMarkdownViewer.js";
-export type {
-  BaseMarkdownViewerProps,
-  MarkdownViewerProps,
-} from "../../markdown-viewer/MarkdownViewerApi.js";
-
-// MarkdownViewer (Media wrapper)
-import { MarkdownViewer as _MarkdownViewer } from "../../markdown-viewer/MarkdownViewer.js";
-import { withOsdkMetrics } from "../../util/withOsdkMetrics.js";
-export const MarkdownViewer: typeof _MarkdownViewer = withOsdkMetrics(
-  _MarkdownViewer,
-  "MarkdownViewer",
-);
