--- packages/react-components/src/public/experimental/tiff-viewer.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// BaseTiffViewer
-export { BaseTiffViewer } from "../../images/tiff-viewer/BaseTiffViewer.js";
-export type {
-  BaseTiffViewerProps,
-  TiffViewerProps,
-} from "../../images/tiff-viewer/TiffViewerApi.js";
-
-// TiffViewer (Media wrapper)
-import { TiffViewer as _TiffViewer } from "../../images/tiff-viewer/TiffViewer.js";
-import { withOsdkMetrics } from "../../util/withOsdkMetrics.js";
-export const TiffViewer: typeof _TiffViewer = withOsdkMetrics(
-  _TiffViewer,
-  "TiffViewer",
-);
